reestructure vectors module
