+++ raphtory/src/vectors/entity_id.rs
@@ -1,0 +1,66 @@
+use crate::{
+    db::graph::{edge::EdgeView, vertex::VertexView},
+    prelude::{EdgeViewOps, GraphViewOps, VertexViewOps},
+};
+use serde::Serializer;
+use std::fmt::{Display, Formatter};
+
+#[derive(Clone, Debug, Eq, PartialEq, Hash)]
+pub(crate) enum EntityId {
+    Node { id: u64 },
+    Edge { src: u64, dst: u64 },
+}
+
+impl EntityId {
+    fn as_node(&self) -> u64 {
+        match self {
+            EntityId::Node { id } => *id,
+            EntityId::Edge { .. } => panic!("edge id unwrapped as a node id"),
+        }
+    }
+}
+
+impl<G: GraphViewOps> From<&VertexView<G>> for EntityId {
+    fn from(value: &VertexView<G>) -> Self {
+        EntityId::Node { id: value.id() }
+    }
+}
+
+impl<G: GraphViewOps> From<VertexView<G>> for EntityId {
+    fn from(value: VertexView<G>) -> Self {
+        EntityId::Node { id: value.id() }
+    }
+}
+
+impl<G: GraphViewOps> From<&EdgeView<G>> for EntityId {
+    fn from(value: &EdgeView<G>) -> Self {
+        EntityId::Edge {
+            src: value.src().id(),
+            dst: value.dst().id(),
+        }
+    }
+}
+
+impl<G: GraphViewOps> From<EdgeView<G>> for EntityId {
+    fn from(value: EdgeView<G>) -> Self {
+        EntityId::Edge {
+            src: value.src().id(),
+            dst: value.dst().id(),
+        }
+    }
+}
+
+impl Display for EntityId {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            EntityId::Node { id } => f.serialize_u64(*id),
+            EntityId::Edge { src, dst } => {
+                f.serialize_u64(*src)
+                    .expect("src ID couldn't be serialized");
+                f.write_str("-")
+                    .expect("edge ID separator couldn't be serialized");
+                f.serialize_u64(*dst)
+            }
+        }
+    }
+}
